Refactored code, added enemy behavior, settings, fixed couple bugs
Refactored code: now the ball bouncing system works through events, so that we can re-use the bounce methods in other events such as the projectile bounce
Worked on enemy base: every enemy will be configured through a scriptable object. There will be a prefab for each enemy type, and the corresponding scriptable object with the settings has to be dragged into the inspector for it to work. There will be different types of enemy behaviors. These inherit from the Enemy class, which contains the base for this.